move files to tools
