+++ docs/api/testing.md
@@ -1,11 +1,6 @@
 # Testing
 
-## Transformers Models
-
 ```{eval-rst}
-.. autofunction:: onnxscript.testing.transformers_models.get_model_and_inputs
-```
-
-```{eval-rst}
-.. autofunction:: onnxscript.testing.transformers_models.phi.get_phi_model_config
+.. automodule:: onnxscript.testing
+    :members:
 ```
