+++ docs/api/tools.md
@@ -1,0 +1,11 @@
+# Tools
+
+## Transformers Models
+
+```{eval-rst}
+.. autofunction:: onnxscript.tools.transformers_models.get_model_and_inputs
+```
+
+```{eval-rst}
+.. autofunction:: onnxscript.tools.transformers_models.phi.get_phi_model_config
+```
