+++ onnxscript/tools/benchmark/benchmark_helpers.py
@@ -471,6 +471,7 @@ def run_inference(
 
 class WrapInferenceSessionForTorch:
     def __init__(self, sess: "onnxruntime.InferenceSession"):  # noqa: F821
+        # onnxruntime is importing when needed as it takes a couple of seconds if it contains CUDA EP.
         import onnxruntime
         import torch
         from onnxruntime.capi import _pybind_state as ORTC  # noqa: N812
@@ -621,6 +622,7 @@ def run_onnx_inference(
         print(f"[run_inference] create session with providers {providers!r}")
 
     begin = time.perf_counter()
+    # onnxruntime is importing when needed as it takes a couple of seconds if it contains CUDA EP.
     import onnxruntime
 
     so = onnxruntime.SessionOptions()
