+++ onnxscript/tools/transformers_models/export_phi_test.py
@@ -9,11 +9,11 @@ import torch
 
 import onnxscript.optimizer
 import onnxscript.rewriter
-import onnxscript.testing.training_helper
-import onnxscript.testing.transformers_models
-import onnxscript.testing.transformers_models.phi
+import onnxscript.tools.training_helper
+import onnxscript.tools.transformers_models
+import onnxscript.tools.transformers_models.phi
 
-HAS_TRANSFORMERS = onnxscript.testing.transformers_models.has_transformers()
+HAS_TRANSFORMERS = onnxscript.tools.transformers_models.has_transformers()
 
 
 def export_to_onnx(model, *input_tensors, optimize=True):
@@ -35,7 +35,7 @@ class TestExportPhi(unittest.TestCase):
     @unittest.skipIf(sys.platform == "win32", reason="not supported yet on Windows")
     @unittest.skipIf(not HAS_TRANSFORMERS, reason="transformers is missing")
     def test_phi_export_cpu(self):
-        model, input_tensors = onnxscript.testing.transformers_models.phi.get_phi_model()
+        model, input_tensors, _ = onnxscript.tools.transformers_models.phi.get_phi_model()
         input_tensors = input_tensors[0]
         expected = model(*input_tensors)
         proto = export_to_onnx(model, *input_tensors)
@@ -52,7 +52,7 @@ class TestExportPhi(unittest.TestCase):
     @unittest.skipIf(not torch.cuda.is_available(), reason="CUDA not available")
     @unittest.skipIf(not HAS_TRANSFORMERS, reason="transformers is missing")
     def test_phi_export_cuda(self):
-        model, input_tensors = onnxscript.testing.transformers_models.phi.get_phi_model()
+        model, input_tensors, _ = onnxscript.tools.transformers_models.phi.get_phi_model()
         input_tensors = input_tensors[0]
         model = model.to("cuda")
         input_tensors = [i.to("cuda") for i in input_tensors]
@@ -70,11 +70,11 @@ class TestExportPhi(unittest.TestCase):
     @unittest.skipIf(sys.platform == "win32", reason="not supported yet on Windows")
     @unittest.skipIf(not HAS_TRANSFORMERS, reason="transformers is missing")
     def test_phi_dort_static(self):
-        model, input_tensors = onnxscript.testing.transformers_models.phi.get_phi_model()
+        model, input_tensors, _ = onnxscript.tools.transformers_models.phi.get_phi_model()
         input_tensors = input_tensors[0]
         expected = model(*input_tensors)
 
-        local_aot_ort = onnxscript.testing.training_helper.make_aot_ort(dynamic=False)
+        local_aot_ort = onnxscript.tools.training_helper.make_aot_ort(dynamic=False)
 
         compiled_model = torch.compile(
             copy.deepcopy(model),
@@ -86,10 +86,10 @@ class TestExportPhi(unittest.TestCase):
         results = compiled_model(*input_tensors)
         torch.testing.assert_allclose(expected[0], results[0], atol=1e-5, rtol=1e-5)
 
-        expected_gradients = onnxscript.testing.training_helper.train_loop(
+        expected_gradients = onnxscript.tools.training_helper.train_loop(
             model, *input_tensors
         )
-        gradients = onnxscript.testing.training_helper.train_loop(
+        gradients = onnxscript.tools.training_helper.train_loop(
             compiled_model, *input_tensors
         )
         torch.testing.assert_allclose(
